docker/open-webui/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
bcrypt==4.1.1